+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Update to add new gague type for Wifi and adjusted the signal strength dBm to percent conversion (thanks @MichaelB2018)
 - Added Nee-vo tank monitor add on (thanks @@momoz)
 - Update for better alarm handing on Power Zone 200
+- Minor update to feedback email so they are always sent as plain text
 
 ## 2.0.01 -2026-05-18
 - Various bug fixes and typo corrections